+++ main.cpp
@@ -7,7 +7,7 @@
 #include "AppWorkerThread.hpp"
 #include "BTReceiver.hpp"
 #include "BTReceiverTCP.hpp"
-#include "BTReceiver/BTReceiverUDP.hpp"
+#include "BTReceiverUDP.hpp"
 #include "BTTransmitter.hpp"
 #include "BTTransmitterTCP.hpp"
 #include "BTTransmitterUDP.hpp"
